# src/build/swagger.json
{
	"components": {
...
		"schemas": {
			"InvalidTokenErrorResponse": {
				"description": "The response given when the api key provided in a request is invalid",
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "Invalid Token"
				}
			},
			"UnauthorizedErrorResponse": {
				"description": "The response given when an api key is not provided in a request that requires it",
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "Access Denied"
				}
			},
			"verifyPinErrorType": {
				"properties": {
					"errorType": {
						"type": "string"
					},
					"errorMessage": {
						"type": "string"
					}
				},
				"required": [
					"errorMessage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"verifyPinResponse": {
				"description": "The response given from a verify pin request.\nIf verified is false, the reason that the PIN was not verified is given.",
				"properties": {
					"verified": {
						"type": "boolean"
					},
					"livePinId": {
						"type": "string"
					},
					"reason": {
						"$ref": "#/components/schemas/verifyPinErrorType"
					}
				},
				"required": [
					"verified"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"verified": false,
					"reason": {
						"errorType": "NotFoundError",
						"errorMessage": "PIN was unable to be verified"
					}
				}
			},
...
			"requiredFieldErrorType": {
				"description": "A error that can occur when you don't give a required input parameter in the body of a request",
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "Must provide an expiration name when expiring a PIN"
				}
			},
			"serverErrorType": {
				"description": "Generic unknown server error",
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "Internal Server Error"
				}
			},
...
			"forbiddenError": {
				"description": "Forbidden Error",
				"properties": {
					"message": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"code"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "Forbidden error",
					"code": 403
				}
			},
...
			"expirationReason": {
				"description": "The reason for expiring a PIN.\n- OptOut = 'OP'\n- CallCenterPinReset = 'CC'\n- OnlineReset = 'OR'\n- ChangeOfOwnership = 'CO'",
				"enum": [
					"OP",
					"CC",
					"OR",
					"CO"
				],
				"type": "string"
			},
...
			"pinRangeErrorType": {
				"description": "Generic range error that may occur when you give a value out of bounds to a pin function",
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "PIN must be of length 1 or greater"
				}
			},
...
			"EntityNotFoundErrorType": {
				"description": "Error that occurs when an entity cannot be found in the database meeting your search criteria",
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "Could not find any entity of type \"ActivePin\" matching: {\n\"where\": {\n\"livePinId\": \"e9bee7c0-de39-47b3-9457-34d32cf6feb4\"\n}\n}"
				}
			},
			"createPinRequestBody": {
				"description": "The request body for a pin creation / recreation request.\nThe address fields given are for a mailing address, which is\nnot necessarily the same as the property address. The property address is required to send the\nGCNotify email / text message to the recipient.\nCertain combinations of fields are required in addition to always required fields:\n- (givenName & lastName_1/lastName_2) OR incorporationNumber\n- phoneNumber AND/OR email\n- requesterName and requesterUsername are required if an employee is requesting the creation,\n rather than self serve",
				"properties": {
					"pinLength": {
						"type": "number",
						"format": "double"
					},
					"allowedChars": {
						"type": "string"
					},
					"numberOfOwners": {
						"type": "number",
						"format": "double"
					},
					"phoneNumber": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"pids": {
						"type": "string"
					},
					"givenName": {
						"type": "string"
					},
					"lastName_1": {
						"type": "string"
					},
					"lastName_2": {
						"type": "string"
					},
					"incorporationNumber": {
						"type": "string"
					},
					"addressLine_1": {
						"type": "string"
					},
					"addressLine_2": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"provinceAbbreviation": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"requesterUsername": {
						"type": "string"
					},
					"propertyAddress": {
						"type": "string"
					}
				},
				"required": [
					"numberOfOwners",
					"pids",
					"lastName_1",
					"addressLine_1",
					"propertyAddress"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"email": "example@test.com",
					"pids": "1234|5678",
					"numberOfOwners": 2,
					"givenName": "Jane",
					"lastName_1": "Smith",
					"lastName_2": "Green",
					"addressLine_1": "123 Main St",
					"addressLine_2": "Unit 12",
					"city": "Vancouver",
					"provinceAbbreviation": "BC",
					"country": "Canada",
					"postalCode": "V1V1V1",
					"propertyAddress": "8765 Willow Way, Chilliwack, BC"
				}
			},
			"updatedPIN": {
				"description": "A PIN updated for a homeowner and its corresponding parcel & database identifiers. Defaults to 8 character length and\nall numbers + lowercase letters as the character set.",
				"properties": {
					"pin": {
						"type": "string"
					},
					"pids": {
						"type": "string"
					},
					"livePinId": {
						"type": "string"
					}
				},
				"required": [
					"pids",
					"livePinId"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"pin": "abcdefgh",
					"pids": "1234",
					"livePinId": "cf430240-e5b6-4224-bd71-a02e098cc6e8"
				}
			},
			"serviceBCCreateRequestBody": {
				"description": "Since these results are verified by a human, much less information is required\nthan the standard create / recreate request. The property address is for sending the\nGCNotify email/ text message, and is not used for verification.",
				"properties": {
					"livePinId": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"propertyAddress": {
						"type": "string"
					},
					"pinLength": {
						"type": "number",
						"format": "double"
					},
					"allowedChars": {
						"type": "string"
					}
				},
				"required": [
					"livePinId",
					"propertyAddress"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"livePinId": "82dc08e5-cbca-40c2-9d35-a4d1407d5f8d",
					"email": "example@test.com",
					"phoneNumber": "+19021234567",
					"propertyAddress": "123 Main Street, Vancouver, BC"
				}
			},
			"PIN": {
				"type": "string",
				"example": "abcdefgh",
				"description": "A PIN generated for a homeowner. Defaults to 8 character length and\nall numbers + lowercase letters as the character set."
			},
			"PINObject": {
				"description": "An object containing an array of unique, generated PINS",
				"properties": {
					"pins": {
						"items": {
							"$ref": "#/components/schemas/PIN"
						},
						"type": "array"
					}
				},
				"required": [
					"pins"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"pins": [
						"abcdefgh",
						"12345678"
					]
				}
			},
			"ActivePin": {
				"properties": {
					"livePinId": {
						"type": "string"
					},
					"pin": {
						"type": "string",
						"nullable": true
					},
					"pids": {
						"type": "string"
					},
					"titleNumber": {
						"type": "string"
					},
					"landTitleDistrict": {
						"type": "string"
					},
					"titleStatus": {
						"type": "string"
					},
					"fromTitleNumber": {
						"type": "string",
						"nullable": true
					},
					"fromLandTitleDistrict": {
						"type": "string",
						"nullable": true
					},
					"givenName": {
						"type": "string",
						"nullable": true
					},
					"lastName_1": {
						"type": "string",
						"nullable": true
					},
					"lastName_2": {
						"type": "string",
						"nullable": true
					},
					"incorporationNumber": {
						"type": "string",
						"nullable": true
					},
					"addressLine_1": {
						"type": "string",
						"nullable": true
					},
					"addressLine_2": {
						"type": "string",
						"nullable": true
					},
					"city": {
						"type": "string",
						"nullable": true
					},
					"provinceAbbreviation": {
						"type": "string",
						"nullable": true
					},
					"provinceLong": {
						"type": "string",
						"nullable": true
					},
					"country": {
						"type": "string",
						"nullable": true
					},
					"postalCode": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"livePinId",
					"pin",
					"pids",
					"titleNumber",
					"landTitleDistrict",
					"titleStatus",
					"fromTitleNumber",
					"fromLandTitleDistrict",
					"givenName",
					"lastName_1",
					"lastName_2",
					"incorporationNumber",
					"addressLine_1",
					"addressLine_2",
					"city",
					"provinceAbbreviation",
					"provinceLong",
					"country",
					"postalCode",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"expireRequestBody": {
				"description": "The request body for a pin expiration request.\nNote that expiredByUsername, propertyAddress, email and phone\nnumber are only required for reasons other\nthan \"CO\" (change of ownership).",
				"properties": {
					"livePinId": {
						"type": "string"
					},
					"expirationReason": {
						"$ref": "#/components/schemas/expirationReason"
					},
					"expiredByUsername": {
						"type": "string"
					},
					"propertyAddress": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"livePinId",
					"expirationReason"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"livePinId": "ca609097-7b4f-49a7-b2e9-efb78afb3ae6",
					"expirationReason": "OP",
					"expiredByUsername": "jsmith",
					"propertyAddress": "123 example st",
					"email": "test@gmail.com"
				}
			},
			"verifyPinRequestBody": {
				"description": "The information needed to verify a pin from the VHERS side.\nNote that the pids are seperated by a vertical bar (|)",
				"properties": {
					"pin": {
						"type": "string"
					},
					"pids": {
						"type": "string"
					}
				},
				"required": [
					"pin",
					"pids"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"pin": "ABCD1234",
					"pids": "12345678|11234567"
				}
			},
...
		},
		"securitySchemes": {
			"vhers_api_key": {
				"type": "apiKey",
				"in": "header",
				"name": "x-api-key",
				"schema": {
					"type": "string",
					"format": "uuid"
				},
				"required": true
			}
		}
	},
	"info": {
		"title": "bc-emli-pin-mgmt-be",
		"version": "1.0.0",
		"description": "BC Virtual Home Energy Rating PIN Management System Backend",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {
			"name": "Government of British Columbia"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/pins/score": {
			"post": {
				"operationId": "AddressScore",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidTokenErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityNotFoundErrorType"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/serverErrorType"
								}
							}
						}
					}
				},
				"description": "Shows off the address match score function without actually updating pins.",
				"security": [
					{
						"vhers_api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/createPinRequestBody"
							}
						}
					}
				}
			}
		},
		"/pins/thresholds": {
			"get": {
				"operationId": "WeightsThresholds",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidTokenErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/serverErrorType"
								}
							}
						}
					}
				},
				"description": "Displays the current weights, thresholds and fuzzy match coefficients used for scoring addresses",
				"security": [
					{
						"vhers_api_key": []
					}
				],
				"parameters": []
			}
		},
		"/pins/vhers-create": {
			"post": {
				"operationId": "CreatePin",
				"responses": {
					"200": {
						"description": "An object containing the unique PIN",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/updatedPIN"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidTokenErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityNotFoundErrorType"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/serverErrorType"
								}
							}
						}
					}
				},
				"description": "Used to create a single, unique PIN, checking against the DB to do so.\nNote that the address line, province, country and postal code information is that of the\nmailing address used for identity verfication. The property address could differ, and will be used\nin the GCNotify email / text message that is sent.\nExpected error codes and messages:\n- `400`\n-- `Invalid Token`\n- `401`\n-- `Access Denied`\n- `422`\n-- `PIN must be of length 1 or greater`\n-- `Too many PIN creation attempts: consider expanding your pin length or character set to allow more unique PINs.`\n-- `Error(s) occured in batchUpdatePin: []`\n- `500`\n -- `Internal Server Error`",
				"security": [
					{
						"vhers_api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/createPinRequestBody"
							}
						}
					}
				}
			}
		},
		"/pins/vhers-regenerate": {
			"post": {
				"operationId": "RecreatePin",
				"responses": {
					"200": {
						"description": "An object containing the unique PIN",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/updatedPIN"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidTokenErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityNotFoundErrorType"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/serverErrorType"
								}
							}
						}
					}
				},
				"description": "Used to recreate a single, unique PIN, checking against the DB to do so.\nNote that the address line, province, country and postal code information is that of the\nmailing address used for identity verfication. The property address could differ, and will be used\nin the GCNotify email / text message that is sent.\nExpected error codes and messages:\n- `422`\n-- `PIN must be of length 1 or greater`\n-- `Too many PIN creation attempts: consider expanding your pin length or character set to allow more unique PINs.`\n-- `Error(s) occured in batchUpdatePin: []`\n- `500`\n -- `Internal Server Error`",
				"security": [
					{
						"vhers_api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/createPinRequestBody"
							}
						}
					}
				}
			}
		},
		"/pins/create": {
			"post": {
				"operationId": "ServiceBCCreatePin",
				"responses": {
					"200": {
						"description": "An object containing the unique PIN",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/updatedPIN"
									},
									"type": "array"
								}
							}
						}
					},
					"403": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/forbiddenError"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityNotFoundErrorType"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/serverErrorType"
								}
							}
						}
					}
				},
				"description": "Used to create a single, unique PIN, checking against the DB to do so.\nThis endpoint has minimal validation, as the validation is expected to be performed by a human.\nNote that the property address given is just used for the GCNotify email / text message\nand is not used for validation.\nExpected error codes and messages:\n- `422`\n-- `PIN must be of length 1 or greater`\n-- `Too many PIN creation attempts: consider expanding your pin length or character set to allow more unique PINs.`\n-- `Error(s) occured in batchUpdatePin: []`\n- `500`\n -- `Internal Server Error`",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/serviceBCCreateRequestBody"
							}
						}
					}
				}
			}
		},
		"/pins/regenerate": {
			"post": {
				"operationId": "ServiceBCRecreatePin",
				"responses": {
					"200": {
						"description": "An object containing the unique PIN",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/updatedPIN"
									},
									"type": "array"
								}
							}
						}
					},
					"403": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/forbiddenError"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityNotFoundErrorType"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/serverErrorType"
								}
							}
						}
					}
				},
				"description": "Used to recreate a single, unique PIN, checking against the DB to do so.\nThis endpoint has minimal validation, as the validation is expected to be performed by a human.\nNote that the property address given is just used for the GCNotify email / text message\nand is not used for validation.\nExpected error codes and messages:\n- `422`\n-- `PIN must be of length 1 or greater`\n-- `Too many PIN creation attempts: consider expanding your pin length or character set to allow more unique PINs.`\n-- `Error(s) occured in batchUpdatePin: []`\n- `500`\n -- `Internal Server Error`",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/serviceBCCreateRequestBody"
							}
						}
					}
				}
			}
		},
		"/pins/initial-create": {
			"get": {
				"operationId": "GetInitialPins",
				"responses": {
					"200": {
						"description": "An object containing an array of the created, unique PINs",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PINObject"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/pinRangeErrorType"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/serverErrorType"
								}
							}
						}
					}
				},
				"description": "Used for the initial creation of PINs, when none exist in the database yet.\nExpected error codes and messages:\n- `422`\n\t-- `PIN must be of length 1 or greater`\n\t-- `Quantity of PINs requested too high: guaranteed repeats for the given pin length and character set.`\n\t-- `The number of PINS created must be greater than 0.`\n- `500`\n\t-- `Internal Server Error`",
				"security": [],
				"parameters": [
					{
						"description": "The quantity of PINs you wish to create.",
						"in": "query",
						"name": "quantity",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The length of each PIN. Defaults to 8 characters.",
						"in": "query",
						"name": "pinLength",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "A string (not regex) of the characters to be used to\ngenerate the pin. Default is A-Z excluding O, and 1-9",
						"in": "query",
						"name": "allowedChars",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/pins/expire": {
			"post": {
				"operationId": "ExpirePin",
				"responses": {
					"200": {
						"description": "The deleted pin",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ActivePin"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/requiredFieldErrorType"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/serverErrorType"
								}
							}
						}
					}
				},
				"description": "Used for expiring pins by their id (livePinId). Requires a reason for expiration, and if not a change of ownership, the name and username of who is expiring the pin.\nExpected error codes and messages:\n- `422`\n\t-- `Could not find any entity of type \"ActivePin\" matching: {\\n    \"livePinId\": \"id here\"\\n}`\n\t-- `Must provide an expiration name when expiring a PIN`\n\t-- `Must provide an expiration username when expiring a PIN`\n- `500`\n\t-- `Internal Server Error`",
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The body of the request. Note that expiredByUsername is only required for reasons other than \"CO\" (change of ownership).",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/expireRequestBody",
								"description": "The body of the request. Note that expiredByUsername is only required for reasons other than \"CO\" (change of ownership)."
							}
						}
					}
				}
			}
		},
		"/pins/etl-expire": {
			"post": {
				"operationId": "ExpirePinEtl",
				"responses": {
					"200": {
						"description": "The deleted pin",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ActivePin"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidTokenErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/requiredFieldErrorType"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/serverErrorType"
								}
							}
						}
					}
				},
				"description": "Used for expiring pins by their id (livePinId) from the etl job. Requires a reason for expiration, and if not a change of ownership, the name and username of who is expiring the pin.\nExpected error codes and messages:\n- `422`\n\t-- `Could not find any entity of type \"ActivePin\" matching: {\\n    \"livePinId\": \"id here\"\\n}`\n\t-- `Must provide an expiration name when expiring a PIN`\n\t-- `Must provide an expiration username when expiring a PIN`\n- `500`\n\t-- `Internal Server Error`",
				"security": [
					{
						"vhers_api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The body of the request. Note that expiredByUsername is only required for reasons other than \"CO\" (change of ownership).",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/expireRequestBody",
								"description": "The body of the request. Note that expiredByUsername is only required for reasons other than \"CO\" (change of ownership)."
							}
						}
					}
				}
			}
		},
		"/pins/verify": {
			"post": {
				"operationId": "VerifyPin",
				"responses": {
					"200": {
						"description": "verified as true if verification was successful, and false otherwise along with a reason",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/verifyPinResponse"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidTokenErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"407": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/verifyPinResponse"
								}
							}
						}
					},
					"408": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/verifyPinResponse"
								}
							}
						}
					},
					"418": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/verifyPinResponse"
								}
							}
						}
					}
				},
				"description": "Verifies the user given a PIN and the pid(s) associated with the title.\nNote: some of the error return codes here are not semantically correct. \nThis is because of external vendor requirements and is unavoidable.",
				"security": [
					{
						"vhers_api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The body for the request. Note that pids should be seperated by a vertical bar (|)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/verifyPinRequestBody",
								"description": "The body for the request. Note that pids should be seperated by a vertical bar (|)"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}